bibox: refactor connection management
Bibox only allows 20 subscriptions per connetion. Bibox also limits the number of
connections that are allowed. The latter point break the usage of
BasicMultiClient because it creates a single socket per subscription which can
be extremely chatty.

This checkin implements a new custom client that pools subscriptions into
20 subscriptions per socket. This also adds subscriptiion throttling, but does
so via timeout. It does not wait for a success response from the server. This
point may cause issues on reconnection and may need to be revisited.