huobi: fixing l2snapshot timestamp
This fix addresses the issue where multiple live sockets connected to huobi
return different timestamps for the same market snapshot data.

Prior to this commit the l2snapshot was using a timestamp generated for
the ws message. The snapshot sent by huobi contains its own timestamp
and version properties (ts and version respectively). The snapshot now
correctly sends the timestamp associated with the actual order book snapshot.
It also now includes a sequenceId that is the version. The version is not
sequentially increasing.: