general: batching utility
This utility adds debounced based batching. It will form the basis for some
flow control utilities that can be used across different implementations. This
initial implementation allows calls to be added to a queue and debounce a
timeout. When the timeout expires a ready event is called with a batch of
items.