general: adding collection/debouncing tool
This class allows us to defer execution of a function until a timeout clears.
The collection of arguments depends on one of two (currently) strategies.

The last strategy will only keep the last arguments and will call our delegate
function with the last arguments.

The batched strategy will keep all arguments and when the timeout executes call
the delegate function with batches of arguments.

These two strategies will support traditional debouncing (via the last
strategy) and a debounced "batching" where we can make a function call
repeatedly and then call the delegate with batches of calls.