kucoin: add sequenceId to l2update and l2snapshot
Change adds sequenceId and sequenceLast to KucoinClient. It removes
timestampMs. It updates the integration tests as well.

Kucoin include similar behavior to Binance in that it includes a sequenceStart
and sequenceEnd values that include the values included in the update. These
are now returned as part of the object so a consumer will have knowledge of
any dropped messages.